+++ CHANGELOG.md
@@ -7,10 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for `--shell` with optional path to an explicit shell as an option
+  when executing `distant spawn` in order to run the command within a shell
+  rather than directly
+
+### Changed
+
+- `distant_protocol::PROTOCOL_VERSION` now uses the crate's major, minor, and
+  patch version at compile-time (parsed via `const-str` crate) to streamline
+  version handling between crate and protocol
+
 ### Fixed
 
 - CLI commands like `distant manager select` will now output errors in a JSON
   format when configured to communicate using JSON
+- `distant-ssh2` no longer caches the remote family globally, but instead
+  caches it per `Ssh` instance
+
+### Removed
+
+- `Cmd::program` and `Cmd::arguments` functions as they were misleading (didn't
+  do what `distant-local` or `distant-ssh2` do)
 
 ## [0.20.0-alpha.12]